[level 0] Title: 구슬을 나누는 경우의 수, Time: 0.01 ms, Memory: 4.02 MB -BaekjoonHub

+++ 프로그래머스/lv0/120840. 구슬을 나누는 경우의 수/구슬을 나누는 경우의 수.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int solution(int balls, int share) {
     long long tp = 1, bt = 1, mn = min(share, balls - share);
-    for (unsigned long long i = 1; i <= mn; ++i) {
+    for (int i = 1; i <= mn; ++i) {
         tp *= balls--;
         if (tp % i == 0) tp /= i;
     }
